Ignore error case for coverage
elasticsearch will not allow client methods to be mocked because it does not expose a reference to the client prototype, and so there is no reproducible way to force the method to throw an error to test handling.

Also, since `semistandard` will already throw an error for an unhandled error case there is no real need to additionally unit test `if (err) return callback(err)` so it can be reasonably safely ignored from coverage reports.